Re organising git code
